go/list: move a generic List into its own package
